% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildZINB-callQTL.R
\name{buildZINB}
\alias{buildZINB}
\title{Build zinb model.}
\usage{
buildZINB(counts)
}
\arguments{
\item{counts}{a vector for gene expression.}
}
\value{
Four parameters
}
\description{
Build zinb model.
}
\examples{
data(testGene)
gene <- unlist(testGene[1, ])
result <-buildZINB(gene)
}
