% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSNPList-callQTL.R
\name{checkSNPList}
\alias{checkSNPList}
\title{Check if the SNP ids in the input genotype matrix are valid.}
\usage{
checkSNPList(snpList, snpDataset)
}
\arguments{
\item{snpList}{a list of SNPs id.}

\item{snpDataset}{SNP dataset chosen from ENSEMBL}
}
\value{
SNP location dataframe
}
\description{
Check if the SNP ids in the input genotype matrix are valid.
}
\examples{
data(testSNP2)
snpList <- rownames(testSNP2)
snpDataset <- 'hsapiens_snp'
snps_loc <- checkSNPList(
  snpList = snpList,
  snpDataset = snpDataset
  )
}
