% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM_normalize-normalizeGene.R
\name{CPM_normalize}
\alias{CPM_normalize}
\title{Normalize the gene expression matrix with CPM.}
\usage{
CPM_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{Input raw gene expression matrix.}
}
\value{
A gene expression matrix after normalized.
}
\description{
`CPM_normalize()` scales an expression matrix using Counts Per Million (CPM)
normalization, applying logarithm and scaling operations to adjust data.
}
\examples{
data(testGene)
CPM_normalize(testGene)
}
