% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGeneLoc-callQTL.R
\name{createGeneLoc}
\alias{createGeneLoc}
\title{Create gene location dataframe}
\usage{
createGeneLoc(geneList, geneDataset)
}
\arguments{
\item{geneList}{a gene id or a list of genes id}

\item{geneDataset}{Gene dataset chosen from the biomart.}
}
\value{
data.frame
}
\description{
Create gene location dataframe
}
\examples{
data(testGene)
geneList <- rownames(testGene)
geneDataset <- 'hsapiens_gene_ensembl'
gene_loc <- createGeneLoc(geneList = geneList,
                          geneDataset = geneDataset)
}
