% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLObject.R
\name{createQTLObject}
\alias{createQTLObject}
\title{createObject:Create the eQTLObject.
We next create a S4 object. The object serves as a container that contains
both data (like the count matrix) and meta.data.}
\usage{
createQTLObject(
  snpMatrix,
  genedata,
  biClassify = FALSE,
  species = NULL,
  group = NULL,
  ...
)
}
\arguments{
\item{snpMatrix}{A genotype matrix where each row is one variant and
each column is one sample, and the scoring method is 0/1/2/3.}

\item{genedata}{A gene expression matrix or a seuratobject.}

\item{biClassify}{The user chooses whether to convert the counting method of
the snpMatrix to 0/1/2, TRUE indicates conversion,
and FALSE indicates no conversion, default is no conversion.}

\item{species}{The species that the user wants to select, human or mouse.}

\item{group}{Provided by Seurat's meta.data, such as celltypes, cellstatus
and so on. By default, it is NULL.}

\item{...}{other parameters}
}
\value{
eQTLObject
}
\description{
createObject:Create the eQTLObject.
We next create a S4 object. The object serves as a container that contains
both data (like the count matrix) and meta.data.
}
\examples{
data(testSNP)
data(testGene)
eqtl <- createQTLObject(snpMatrix = testSNP,
                     genedata = testGene,
                     biClassify = FALSE,
                     species = 'human',
                     group = NULL)

}
