% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSNPsLoc-callQTL.R
\name{createSNPsLoc}
\alias{createSNPsLoc}
\title{Create SNP location dataframe}
\usage{
createSNPsLoc(snpList, snpDataset)
}
\arguments{
\item{snpList}{a list of SNPs id.}

\item{snpDataset}{SNP dataset chosen from ENSEMBL}
}
\value{
data.frame
}
\description{
Create SNP location dataframe
}
\examples{
snpList <- c('rs546', 'rs549', 'rs568', 'rs665', 'rs672')
snpDataset <- 'hsapiens_snp'
snp_loc <- createSNPsLoc(snpList = snpList,
                         snpDataset = snpDataset)
}
