% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq_normalize-normalizeGene.R
\name{DESeq_normalize}
\alias{DESeq_normalize}
\title{Normalize the gene expression matrix with DESeq.}
\usage{
DESeq_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{Input raw gene expression matrix.}
}
\value{
A gene expression matrix after normalized.
}
\description{
`DESeq_normalize()`normalizes an expression matrix using the DESeq2 package.
}
\examples{
data(testGene)
DESeq_normalize(testGene)
}
