% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_histplot-visualizeQTL.R
\name{draw_histplot}
\alias{draw_histplot}
\title{Generate a hist plot of expression levels by SNP factor.}
\usage{
draw_histplot(df, unique_group)
}
\arguments{
\item{df}{Data frames listed as gene expression data, genotype data,
and groups}

\item{unique_group}{name of unique group}
}
\value{
ggplot
}
\description{
`draw_histplot()` generates histograms using ggplot2, displaying the
distribution of expression values categorized by SNP type.
}
\examples{
set.seed(123)
counts_Ref <- rnorm(50, mean = 10, sd = 2)
counts_Alt <- rnorm(50, mean = 12, sd = 2)
i <- rep("GroupA", 100);unique_group <- unique(i)
dataframe <- data.frame(expression = c(counts_Ref, counts_Alt),
                        snp = c(rep("REF", length(counts_Ref)),
                        rep("ALT", length(counts_Alt))), group = i)
dataframe$snp <- factor(dataframe$snp, levels = c("REF", "ALT"))
draw_histplot(df = dataframe, unique_group = unique_group)
}
