% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_QTLplot-visualizeQTL.R
\name{draw_QTLplot}
\alias{draw_QTLplot}
\title{Create a combined plot with violin, boxplot, and scatter point overlay.}
\usage{
draw_QTLplot(df, unique_group)
}
\arguments{
\item{df}{Data frames listed as gene expression data, genotype data,
and groups}

\item{unique_group}{name of unique group}
}
\value{
ggplot
}
\description{
`draw_QTLplot()` generates a combined plot using ggplot2, showing the
distribution of expression values across different SNPs. It combines a
violin plot, boxplot, and scatter points for each SNP category.
}
\examples{
set.seed(123)
counts_Ref <- rnorm(50, mean = 10, sd = 2)
counts_Alt <- rnorm(50, mean = 12, sd = 2)
i <- rep("GroupA", 100);unique_group <- unique(i)
dataframe <- data.frame(expression = c(counts_Ref, counts_Alt),
                        snp = c(rep("REF", length(counts_Ref)),
                        rep("ALT", length(counts_Alt))), group = i)
dataframe$snp <- factor(dataframe$snp, levels = c("REF", "ALT"))
draw_QTLplot(df = dataframe, unique_group = unique_group)
}
