% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_violinplot-visualizeQTL.R
\name{draw_violinplot}
\alias{draw_violinplot}
\title{Generate a violin plot of expression levels by SNP factor}
\usage{
draw_violinplot(df, unique_group)
}
\arguments{
\item{df}{Data frames listed as gene expression data, genotype data,
and groups}

\item{unique_group}{name of unique group}
}
\value{
ggplot
}
\description{
`draw_violinplot()` creates a violin plot visualizing expression levels
across different SNP factors in the dataframe. It uses ggplot2 to produce a
plot with customizable aesthetics for clarity and presentation.
}
\examples{
set.seed(123)
counts_Ref <- rnorm(50, mean = 10, sd = 2)
counts_Alt <- rnorm(50, mean = 12, sd = 2)
i <- rep("GroupA", 100);unique_group <- unique(i)
dataframe <- data.frame(expression = c(counts_Ref, counts_Alt),
                        snp = c(rep("REF", length(counts_Ref)),
                        rep("ALT", length(counts_Alt))), group = i)
dataframe$snp <- factor(dataframe$snp, levels = c("REF", "ALT"))
draw_violinplot(df = dataframe, unique_group = unique_group)
}
