% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLObject.R
\docType{class}
\name{eQTLObject-class}
\alias{eQTLObject-class}
\title{Class 'eQTLObject'
The eQTLObject class is an R object designed to store data related to eQTL
analysis, encompassing data lists, result data frames, and layers for
biClassify, species, and grouping information.}
\value{
eQTLObject
}
\description{
Class 'eQTLObject'
The eQTLObject class is an R object designed to store data related to eQTL
analysis, encompassing data lists, result data frames, and layers for
biClassify, species, and grouping information.
}
\section{Slots}{

\describe{
\item{\code{rawData}}{A gene expression dataframe, the row names represent gene IDs
and the column names represent cell IDs.}

\item{\code{filterData}}{list. gene expression matrix after normalizing.}

\item{\code{eQTLResult}}{The result dataframe obtained the sc-eQTL results.}

\item{\code{biClassify}}{The user chooses whether to convert the counting method of
the snpMatrix to 0, 1, 2, TRUE indicates conversion, and FALSE indicates no
conversion, default is no conversion.}

\item{\code{species}}{The species that the user wants to select, human or mouse.}

\item{\code{groupBy}}{Options for cell grouping, users can choose celltype,
cellstatus, etc., depending on metadata.}

\item{\code{useModel}}{model for fitting dataframe.}
}}

