% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeneSNP.R
\name{filterGeneSNP}
\alias{filterGeneSNP}
\title{filterGeneSNP: Filter gene expression matrix and genotype matrix.}
\usage{
filterGeneSNP(
  eQTLObject,
  snp.number.of.cells.percent = 2,
  expression.min = 0,
  expression.number.of.cells.percent = 2
)
}
\arguments{
\item{eQTLObject}{An S4 object of class eQTLObject.}

\item{snp.number.of.cells.percent}{Only SNPs where cells with each of the
different genotypes (REF and ALT, or AA, Aa, and aa) individually account for
 at least `snp.number.of.cells.percent`% of the total cells will be considered.
 The default valus is 10.}

\item{expression.min}{threshold for valid gene expression levels, utilized
alongside another parameter, expression.number.of.cells. The default value is 0.}

\item{expression.number.of.cells.percent}{Only genes with expression levels
exceeding `expression.min` in at least `expression.number.of.cells.percent`%
of cells are considered. The default value is 10.}
}
\value{
filtered genotype matrix and gene expression matrix.
}
\description{
filterGeneSNP: Filter gene expression matrix and genotype matrix.
}
\examples{
data(testSNP)
data(testGene)
eqtl <- createQTLObject(snpMatrix = testSNP,
                     genedata = testGene,
                     biClassify = FALSE,
                     species = 'human',
                     group = NULL)
eqtl <- normalizeGene(eqtl, method = "logNormalize")
eqtl <- filterGeneSNP(eqtl,
                      snp.number.of.cells.percent = 2,
                      expression.min = 0,
                      expression.number.of.cells.percent = 2
                      )
}
