% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_normalize-normalizeGene.R
\name{limma_normalize}
\alias{limma_normalize}
\title{Normalize the gene expression matrix with limma}
\usage{
limma_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{Input raw gene expression matrix.}
}
\value{
A gene expression matrix after normalized.
}
\description{
`limma_normalize()`normalizes an expression matrix using the quantile
normalization method provided by the limma package.
}
\examples{
data(testGene)
limma_normalize(testGene)
}
