% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModel-utils.R
\name{linearModel}
\alias{linearModel}
\title{Linear model fitting the gene expression matrix and genotype matrix.}
\usage{
linearModel(
  eQTLObject,
  geneIDs,
  snpIDs,
  biClassify = FALSE,
  pAdjustMethod = "bonferroni",
  pAdjustThreshold = 0.05,
  logfcThreshold = 0.1
)
}
\arguments{
\item{eQTLObject}{An S4 object of class eQTLObject.}

\item{geneIDs}{Matching genes can be used to fit data.}

\item{snpIDs}{Matching SNPs can be used to fit data.}

\item{biClassify}{The user chooses whether to convert the counting method of
the snpMatrix to 0/1/2, TRUE indicates conversion, and FALSE indicates no
conversion, default is no conversion.}

\item{pAdjustMethod}{Methods for p-value adjusting, one of "bonferroni",
"holm", "hochberg", "hommel" or "BH". The default option is "bonferroni".}

\item{pAdjustThreshold}{Only SNP-Gene pairs with adjusted p-values meeting
the threshold will be displayed. Default by 0.05.}

\item{logfcThreshold}{Represents the minimum beta threshold for fitting
SNP-Gene pairs. Default by 0.1.}
}
\value{
Dataframe that contains gene-SNP pairs' information.
}
\description{
Linear model fitting the gene expression matrix and genotype matrix.
}
\examples{
data(testEQTL)
Gene <- rownames(slot(testEQTL, "filterData")$normExpMat)
SNP <- rownames(slot(testEQTL, "filterData")$snpMat)
linearResult <- linearModel(
  eQTLObject = testEQTL,
  geneIDs = Gene,
  snpIDs = SNP,
  biClassify = FALSE,
  pAdjustMethod = "bonferroni",
  pAdjustThreshold = 0.05,
  logfcThreshold = 0.025)
}
