% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normalize-normalizeGene.R
\name{log_normalize}
\alias{log_normalize}
\title{Normalize the gene expression matrix with logNormalize method.}
\usage{
log_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{Input raw gene expression matrix.}
}
\value{
A gene expression matrix after normalized.
}
\description{
`log_normalize()` transforms an expression matrix by applying logarithm and
scaling operations to normalize data.
}
\examples{
data(testGene)
log_normalize(testGene)
}
