% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGene.R
\name{normalizeGene}
\alias{normalizeGene}
\title{normalizeGene: Normalize the gene expression data.}
\usage{
normalizeGene(eQTLObject, method = "logNormalize")
}
\arguments{
\item{eQTLObject}{An S4 object of class eQTLObject.}

\item{method}{Method for normalizing for gene expression dataframe, one of
"logNormalize", "CPM", "TPM", "DESeq" or "limma"}
}
\value{
A normalized gene expression matrix.
}
\description{
Gene expression matrix normalization is necessary to eliminate
technical biases and variabilities, ensuring accurate and comparable
analysis of gene expression data. Here we provide `normalizeGene()`to
normalize the data.
}
\examples{
data(testEQTL)
eqtl <- normalizeGene(testEQTL, method = "logNormalize")
}
