% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonModel-utils.R
\name{poissonModel}
\alias{poissonModel}
\title{Poisson model fitting the gene expression matrix and genotype matrix.}
\usage{
poissonModel(
  eQTLObject,
  geneIDs,
  snpIDs,
  biClassify = FALSE,
  pAdjustMethod = "bonferroni",
  pAdjustThreshold = 0.05,
  logfcThreshold = 0.1
)
}
\arguments{
\item{eQTLObject}{An S4 object of class eQTLObject.}

\item{geneIDs}{Matching genes can be used to fit data.}

\item{snpIDs}{Matching SNPs can be used to fit data.}

\item{biClassify}{The user chooses whether to convert the counting method of
the snpMatrix to 0/1/2, TRUE indicates conversion, and FALSE indicates no
conversion, default is FALSE.}

\item{pAdjustMethod}{Methods for p-value adjusting, one of "bonferroni",
"holm", "hochberg", "hommel" or "BH". The default option is "bonferroni".}

\item{pAdjustThreshold}{Only SNP-Gene pairs with adjusted p-values meeting
the threshold will be displayed. The default value is 0.05.}

\item{logfcThreshold}{Represents the minimum beta threshold for fitting
SNP-Gene pairs.}
}
\value{
Dataframe that contains gene-SNP pairs' information.
}
\description{
Poisson model fitting the gene expression matrix and genotype matrix.
}
\examples{
data(testEQTL)
Gene <- rownames(slot(testEQTL, "filterData")$normExpMat)
SNP <- rownames(slot(testEQTL, "filterData")$snpMat)
poissonResult <- poissonModel(
  eQTLObject = testEQTL,
  geneIDs = Gene,
  snpIDs = SNP,
  biClassify = FALSE,
  pAdjustMethod = "bonferroni",
  pAdjustThreshold = 0.05,
  logfcThreshold = 0.025
)
}
