% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers-visualizeQTL.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from gene expression data and update cell lists}
\usage{
remove_outliers(exprsMat, Geneid, A_cells, B_cells, C_cells = NULL)
}
\arguments{
\item{exprsMat}{Input gene expression matrix}

\item{Geneid}{Chosen gene id.}

\item{A_cells}{A genotype cells}

\item{B_cells}{B genotype cells}

\item{C_cells}{C genotype cells}
}
\value{
a list of cells ids
}
\description{
remove_outliers() is a function designed to process gene expression data
stored in an expression matrix. It identifies outliers within the data based
on the MAD method and filters them out. The function updates specified cell
lists by retaining only those cells that have non-outlier expression values
for a specified gene.
}
\examples{
## Mock expression matrix
set.seed(123)
exprsMat <- matrix(rnorm(200), nrow = 5)
rownames(exprsMat) <- paste0("Gene", 1:nrow(exprsMat))
colnames(exprsMat)  <- paste0("cell", 1:ncol(exprsMat))
A_cells <- colnames(exprsMat)[1:13] # Example A cell list
B_cells <- colnames(exprsMat)[14:26]  # Example B cell list
C_cells <- colnames(exprsMat)[27:40]  # Example C cell list
remove_outliers(exprsMat, "Gene1", A_cells, B_cells, C_cells)
}
