% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQTLtools.R
\docType{package}
\name{scQTLtools-package}
\alias{scQTLtools}
\alias{scQTLtools-package}
\title{scQTLtools: An R package for single-cell eQTL analysis and visualization}
\description{
This package specializes in analyzing and visualizing eQTL at the single-cell level. It can read gene expression matrices or Seurat data, along with genotype data. It offers a function for cis-eQTL analysis to detect eQTL within a given range, and another function to fit models with three methods. Using this package, users can also generate single-cell level visualization result.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/XFWuCN/scQTLtools/issues}
}

}
\author{
\strong{Maintainer}: Xiaofeng Wu \email{1427972815@qq.com} (\href{https://orcid.org/0009-0003-6254-5575}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Xin Huang [copyright holder]
  \item Siwen Xu \email{siwxu@gdpu.edu.cn} [copyright holder]
}

Other contributors:
\itemize{
  \item Jingtong Kang \email{1203178107@qq.com} [compiler]
}

}
\keyword{internal}
