% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLObject.R
\name{show,eQTLObject-method}
\alias{show,eQTLObject-method}
\title{Show Method for eQTLObject Class}
\usage{
\S4method{show}{eQTLObject}(object)
}
\arguments{
\item{object}{An S4 object of class eQTLObject.}
}
\value{
information of eQTLObject
}
\description{
This method is to display information about an object of class eQTLObject.
When called on an eQTLObject, it prints a descriptive message to the console
}
\examples{
data(testEQTL)
testEQTL
}
