% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testEQTL}
\alias{testEQTL}
\title{Test eqtl object}
\format{
A simple \code{object}.
}
\usage{
data(testEQTL)
}
\description{
An `eqtlObject` created by the `createQTLObject` function, where the raw
expression matrix is normalized using `normalizeGene()`, and both the
genotype matrix and the normalized gene expression matrix are filtered
by `filterGeneSNP()`.
}
\examples{
data(testEQTL)
}
\keyword{datasets}
