% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testGene}
\alias{testGene}
\title{Test Gene Expression Dataset}
\format{
A simple \code{matrix}.
}
\usage{
data(testGene)
}
\description{
A dataset containing example gene expression data for testing purposes.
100 rows and 2705 colnums. The row names represent gene IDs or SYMBOL and
the column names represent cell IDs.
}
\examples{
data(testGene)
}
\keyword{datasets}
