% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testSeurat}
\alias{testSeurat}
\title{Test SeuratObject}
\format{
A \code{object}
}
\usage{
data(testSeurat)
}
\description{
A Seurat object for single-cell RNA-seq data.
}
\examples{
data(testSeurat)
}
\keyword{datasets}
