% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testSNP}
\alias{testSNP}
\title{Test Genotype Dataset}
\format{
A simple \code{matrix}.
}
\usage{
data(testSNP)
}
\description{
A dataset containing single nucleotide variant data. 1000 rows and 2705 colnums.
Each row is one variant and each column is one cell.
}
\examples{
data(testSNP)
}
\keyword{datasets}
