% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testSNP2}
\alias{testSNP2}
\title{Test Genotype Dataset}
\format{
A simple \code{matrix}.
}
\usage{
data(testSNP2)
}
\description{
A dataset containing single nucleotide variant data.500 rows and 500 colnums
Each row is one variant and each column is one cell.
}
\examples{
data(testSNP2)
}
\keyword{datasets}
