% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPM_normalize-normalizeGene.R
\name{TPM_normalize}
\alias{TPM_normalize}
\title{Normalize the gene expression matrix with TPM}
\usage{
TPM_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{Input raw gene expression matrix.}
}
\value{
A gene expression matrix after normalized.
}
\description{
`TPM_normalize()` scales an expression matrix using Transcripts Per Million
(TPM) normalization, applying logarithm and scaling operations to adjust
data based on library size.
}
\examples{
data(testGene)
TPM_normalize(testGene)
}
