% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callQTL.R
\name{callQTL}
\alias{callQTL}
\title{callQTL: Uncover single-cell eQTLs exclusively using scRNA-seq data.
A function designed to identify eQTLs from scRNA-seq data.}
\usage{
callQTL(
  eQTLObject,
  gene_ids = NULL,
  downstream = NULL,
  upstream = NULL,
  pAdjustMethod = "bonferroni",
  useModel = "zinb",
  pAdjustThreshold = 0.05,
  logfcThreshold = 0.1
)
}
\arguments{
\item{eQTLObject}{An S4 object of class eQTLObject.}

\item{gene_ids}{A gene ID or a list of gene IDS.}

\item{downstream}{Being used to match SNPs within a base range defined by
the start position of genes.}

\item{upstream}{Being used to match SNPs within a base range defined by the
end position of genes.}

\item{pAdjustMethod}{Methods for p-value adjusting, one of 'bonferroni',
'holm', 'hochberg', 'hommel' or 'BH'. Default by 'bonferroni'.}

\item{useModel}{Model for fitting dataframe, one of 'possion', 'zinb', or
'linear'.}

\item{pAdjustThreshold}{Only SNP-Gene pairs with adjusted p-values meeting
the threshold will be displayed. Default by 0.05.}

\item{logfcThreshold}{Represents the minimum beta threshold for fitting
SNP-Gene pairs.}
}
\value{
A dataframe, each row describes eQTL discovering result of a
SNP-Gene pair.
}
\description{
callQTL: Uncover single-cell eQTLs exclusively using scRNA-seq data.
A function designed to identify eQTLs from scRNA-seq data.
}
\examples{
data(testEQTL)
eqtl <- callQTL(
  eQTLObject = testEQTL,
  gene_ids = NULL,
  downstream = NULL,
  upstream = NULL,
  pAdjustMethod = 'bonferroni',
  useModel = 'linear',
  pAdjustThreshold = 0.05,
  logfcThreshold = 0.025
)
}
