% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeneSNP.R
\name{filterGeneSNP}
\alias{filterGeneSNP}
\title{filterGeneSNP: Filter gene expression matrix and genotype matrix.}
\usage{
filterGeneSNP(
  eQTLObject,
  snpNumOfCellsPercent = 2,
  expressionMin = 0,
  expressionNumOfCellsPercent = 2
)
}
\arguments{
\item{eQTLObject}{An S4 object of class eQTLObject.}

\item{snpNumOfCellsPercent}{Only SNPs where cells with each of the
different genotypes (REF and ALT, or AA, Aa, and aa) individually account
for at least `snpNumOfCellsPercent`% of the total cells will be considered.
Default by 10.}

\item{expressionMin}{threshold for valid gene expression levels, utilized
alongside another parameter,expression.number.of.cells.Default by 0.}

\item{expressionNumOfCellsPercent}{Only genes with expression levels
exceeding `expressionMin` in at least `expressionNumOfCellsPercent`%
of cells are considered. The default value is 10.}
}
\value{
filtered matrices.
}
\description{
filterGeneSNP: Filter gene expression matrix and genotype matrix.
}
\examples{
data(testSNP)
data(testGene)
eqtl <- createQTLObject(snpMatrix = testSNP, genedata = testGene)
eqtl <- normalizeGene(eqtl)
eqtl <- filterGeneSNP(eqtl,
  snpNumOfCellsPercent = 2,
  expressionMin = 0,
  expressionNumOfCellsPercent = 2)
}
