% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeQTL.R
\name{visualizeQTL}
\alias{visualizeQTL}
\title{visualizeQTL: Visualize the gene-snp pairs by group.}
\usage{
visualizeQTL(
  eQTLObject,
  SNPid,
  Geneid,
  groupName = NULL,
  plottype = "QTLplot",
  removeoutlier = FALSE
)
}
\arguments{
\item{eQTLObject}{An S4 object of class eQTLObject.}

\item{SNPid}{ID of SNP.}

\item{Geneid}{ID of Gene.}

\item{groupName}{Users can choose one or more than one single cell groups.}

\item{plottype}{Types of plot,one of "QTLplot","violin","boxplot" or
"histplot".}

\item{removeoutlier}{Whether identify and remove the outliers.
Default by FALSE.}
}
\value{
list
}
\description{
visualizeQTL: Visualize the gene-snp pairs by group.
}
\examples{
data(testEQTL)
## We have to call the eQTLs firstly using `callQTL()`.
eqtl <- callQTL(eQTLObject = testEQTL, useModel = "linear")
visualizeQTL(eQTLObject = eqtl,
SNPid = "1:632647",
Geneid = "RPS27",
groupName = NULL,
plottype = "QTLplot",
removeoutlier = FALSE)
}
